+++ Source/GameFiles/ClientSide/GuiManager.h
@@ -24,7 +24,7 @@ public:
 	void ToggleInventoryGui();
 	void HideInventoryGui();
 	void AddInventoryItemToGui(const Gui_Item_Data gid);
-	void RemoveInventoryItemFromGui(const int ID);
+	void RemoveInventoryItemFromGui(const long ID);
 	void ShowCircularItemGui();
 	void HideCircularItemGui();
 	void ShowLootingGui(std::vector<Gui_Item_Data> gui_Item_Data);
