+++ Source/GameFiles/ClientSide/WorldObject.h
@@ -11,6 +11,7 @@ class WorldObject
 public:
 	WorldObject(const unsigned int id){this->zMesh = 0; this->zID = id;}
 	virtual ~WorldObject(){ if (this->zMesh){ this->zMesh = 0; } }
+
 	/*!	Returns Pointer to the Player Mesh*/
 	iMesh* GetMesh() const {return this->zMesh;}
 	/*! Returns Object Model Scale*/
@@ -42,5 +43,5 @@ public:
 
 protected:
 	iMesh* zMesh;
-	unsigned int zID;
+	long zID;
 };
