+++ Source/GameFiles/Items/Material.cpp
@@ -1,6 +1,6 @@
 #include "Material.h"
 
-Material::Material(const unsigned int id, const unsigned int itemType, const unsigned int craftingType, const int stacksRequiredToCraft)
+Material::Material(const long id, const unsigned int itemType, const unsigned int craftingType, const int stacksRequiredToCraft)
 	: Item(id, itemType)
 {
 	this->zStacking = true;
